pyproject.toml
[tool.poetry]
name = "etl-mongodb"
version = "0.1.0"
description = ""
authors = ["Alexandre Andreo <alexandre.andreo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
pymongo = "^4.11.1"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
